Replace basic example with new syntax example
