--- assets/strings/test/en-US.ftl
@@ -1,1 +1,0 @@
-hello = Hello, world!
